backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.6.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pydantic[email]==2.6.4
pydantic>=2.0.0
pymongo
python-jose[cryptography]
python-dotenv
passlib[bcrypt]
motor
requests
python-multipart
pyjwt==2.8.0
google-generativeai==0.7.2